Delete controllers file
