+++ app/api/routes.go
@@ -1,8 +1,26 @@
 package api
 
 import (
+	alertUsecases "github.com/PaoGRodrigues/tfi-backend/app/usecase/alert"
+	hostUsecases "github.com/PaoGRodrigues/tfi-backend/app/usecase/host"
+	notificationChannelUseCases "github.com/PaoGRodrigues/tfi-backend/app/usecase/notificationchannel"
+	trafficUsecases "github.com/PaoGRodrigues/tfi-backend/app/usecase/traffic"
 	"github.com/gin-gonic/gin"
 )
+
+type Api struct {
+	TrafficSearcher                      *trafficUsecases.GetTrafficFlowsUseCase
+	GetLocalhostsUseCase                 *hostUsecases.GetLocalhostsUseCase
+	GetTrafficFlowsPerDestinationUseCase *trafficUsecases.GetTrafficFlowsPerDestinationUseCase
+	GetTrafficFlowsPerCountryUseCase     *trafficUsecases.GetTrafficFlowsPerCountryUseCase
+	StoreTrafficFlowsUseCase             *trafficUsecases.StoreTrafficFlowsUseCase
+	GetAlertsUseCase                     *alertUsecases.GetAlertsUseCase
+	BlockHostUseCase                     *hostUsecases.BlockHostUseCase
+	ConfigureNotificationChannelUseCase  *notificationChannelUseCases.ConfigureChannelUseCase
+	NotifyAlertsUseCase                  *alertUsecases.NotifyAlertsUseCase
+	StoreHostsUseCase                    *hostUsecases.StoreHostUseCase
+	*gin.Engine
+}
 
 func (api *Api) MapURLToPing() {
 	api.GET("/ping", func(c *gin.Context) {
@@ -49,5 +67,5 @@ func (api *Api) MapStoreHostsURL() {
 }
 
 func (api *Api) MapStoreActiveFlowsURL() {
-	api.POST("/activeflows", api.StoreActiveTraffic)
+	api.POST("/activeflows", api.StoreTrafficFlows)
 }
